update the feedback with snack bar for insertion and deletion from and to DB
